# packages/credential-showcase-openapi/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Credential Showcase API
  version: 0.1.1
servers:
  - url: https://virtserver.swaggerhub.com/SphereonInt/showcase/0.1.0/v1
    description: SwaggerHub API Auto Mocking
  - url: https://api.example.com/v1
...
security:
  - OAuth2: []
  - OIDC: []
paths:
...
  /assets:
    get:
      tags:
        - Assets
      summary: List assets
      operationId: listAssets
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Assets
      summary: Create an asset
      operationId: createAsset
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /assets/{assetId}:
    parameters:
      - name: assetId
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Assets
      summary: Get a specific asset
      operationId: getAsset
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Assets
      summary: Update an asset
      operationId: updateAsset
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetRequest'
        required: true
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Assets
      summary: Delete an asset
      operationId: deleteAsset
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: <https://auth.example.com/oauth/authorize>
          tokenUrl: <https://auth.example.com/oauth/token>
          scopes:
            editor: Grants editor access
            admin: Grants admin access
            viewer: Grants viewer access
    OIDC:
      type: openIdConnect
      openIdConnectUrl: <https://auth.example.com/.well-known/openid-configuration>
      description: OpenID Connect security scheme for OAuth2 flows
  schemas:
    Asset:
      type: object
      required:
        - id
        - mediaType
        - content
      properties:
        id:
          type: string
          description: Unique identifier for the asset in the database
          example: 123e4567-e89b-12d3-a456-426614174469
        mediaType:
          type: string
          description: Type of asset (e.g., image, document, icon)
          example: image/jpeg
        content:
          type: string
          description: Binary data of the asset
          example: base64 encoded binary data
        fileName:
          type: string
          description: Name of the asset file
          example: asset.jpg
        description:
          type: string
          description: Description of the asset
          example: A beautiful image of a cat
    AssetRequest:
      type: object
      required:
        - mediaType
        - content
      properties:
        mediaType:
          type: string
          description: Type of asset (e.g., image, document, icon)
          example: image/jpeg
        content:
          type: string
          description: Binary data of the asset
          example: base64 encoded binary data
        fileName:
          type: string
          description: Name of the asset file
          example: asset.jpg
        description:
          type: string
          description: Description of the asset
          example: A beautiful image of a cat
    AssetsResponse:
      type: object
      required:
        - assets
      properties:
        assets:
          type: array
          items:
            $ref: '#/components/schemas/Asset'
    AssetResponse:
      type: object
      required:
        - asset
      properties:
        asset:
          $ref: '#/components/schemas/Asset'
...
  responses:
    'BadRequest':
      description: Bad Request
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Invalid request data'
    'Unauthorized':
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Invalid authentication credentials'
    'Forbidden':
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Insufficient permissions to access resource'
...
    'InternalServerError':
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'An unexpected error occurred'